refactor: extract PHP bootstrap helpers
